requirements.txt
aiohttp
aiodns
pyyaml
rich
jinja2
beautifulsoup4
tqdm
dnspython
click
